moved pointer creation to a generic place
