+++ Assets/MixedRealityToolkit/_Core/Devices/OpenVR/OpenVRDeviceManager.cs
@@ -205,67 +205,32 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Devices.OpenVR
                     return null;
             }
 
-            var pointers = new List<IMixedRealityPointer>();
+            var pointers = RequestPointers(controllerType, controllingHand);
+            var inputSource = InputSystem?.RequestNewGenericInputSource($"{CurrentControllerType} Controller {controllingHand}", pointers);
 
-            if (MixedRealityManager.HasActiveProfile &&
-                MixedRealityManager.Instance.ActiveProfile.IsInputSystemEnabled &&
-                MixedRealityManager.Instance.ActiveProfile.PointerProfile != null)
-            {
-                for (int i = 0; i < MixedRealityManager.Instance.ActiveProfile.PointerProfile.PointerOptions.Length; i++)
-                {
-                    var pointerProfile = MixedRealityManager.Instance.ActiveProfile.PointerProfile.PointerOptions[i];
-
-                    if (pointerProfile.ControllerType.Type == null ||
-                        pointerProfile.ControllerType == controllerType)
-                    {
-                        if (pointerProfile.Handedness == Handedness.Any ||
-                            pointerProfile.Handedness == Handedness.Both ||
-                            pointerProfile.Handedness == controllingHand)
-                        {
-                            var pointerObject = UnityEngine.Object.Instantiate(pointerProfile.PointerPrefab);
-                            var pointer = pointerObject.GetComponent<IMixedRealityPointer>();
-
-                            if (pointer != null)
-                            {
-                                pointers.Add(pointer);
-                            }
-                            else
-                            {
-                                Debug.LogWarning($"Failed to attach {pointerProfile.PointerPrefab.name} to {controllerType.Name}.");
-                            }
-                        }
-                    }
-                }
-            }
-
-            IMixedRealityInputSource inputSource = InputSystem?.RequestNewGenericInputSource($"{CurrentControllerType} Controller {controllingHand}", pointers.ToArray());
             GenericOpenVRController detectedController = null;
 
             switch (CurrentControllerType)
             {
                 case SupportedControllerType.GenericOpenVR:
                     detectedController = new GenericOpenVRController(TrackingState.NotTracked, controllingHand, inputSource);
-                    detectedController.SetupConfiguration(typeof(GenericOpenVRController));
                     break;
                 case SupportedControllerType.ViveWand:
                     detectedController = new ViveWandController(TrackingState.NotTracked, controllingHand, inputSource);
-                    detectedController.SetupConfiguration(typeof(ViveWandController));
                     break;
                 case SupportedControllerType.ViveKnuckles:
                     detectedController = new ViveKnucklesController(TrackingState.NotTracked, controllingHand, inputSource);
-                    detectedController.SetupConfiguration(typeof(ViveKnucklesController));
                     break;
                 case SupportedControllerType.OculusTouch:
                     detectedController = new OculusTouchController(TrackingState.NotTracked, controllingHand, inputSource);
-                    detectedController.SetupConfiguration(typeof(OculusTouchController));
                     break;
                 case SupportedControllerType.OculusRemote:
                     detectedController = new OculusRemoteController(TrackingState.NotTracked, controllingHand, inputSource);
-                    detectedController.SetupConfiguration(typeof(OculusRemoteController));
                     break;
             }
 
             Debug.Assert(detectedController != null);
+            detectedController?.SetupConfiguration(controllerType);
 
             for (int i = 0; i < detectedController?.InputSource?.Pointers?.Length; i++)
             {
